docs: plugin options refactor: move to core file fix: umami script condition wip: website using plugin
