+++ packages/unplugin-panda-macro/playground/vite.config.ts
@@ -3,5 +3,5 @@ import Inspect from 'vite-plugin-inspect'
 import panda from '../src/vite'
 
 export default defineConfig({
-  plugins: [Inspect(), panda({ output: 'grouped' })],
+  plugins: [Inspect(), panda({ output: 'grouped', outfile: './panda.css' })],
 })
