+++ packages/unplugin-panda-macro/src/rollup.ts
@@ -1,4 +1,4 @@
 import { createRollupPlugin } from 'unplugin'
-import { unpluginFactory } from '.'
+import { unpluginFactory } from './plugin/core'
 
 export default createRollupPlugin(unpluginFactory)
