--- packages/unplugin-panda-macro/src/types.ts
@@ -1,3 +1,0 @@
-export interface Options {
-  // define your plugin options here
-}
