+++ packages/unplugin-panda-macro/src/webpack.ts
@@ -1,4 +1,4 @@
 import { createWebpackPlugin } from 'unplugin'
-import { unpluginFactory } from '.'
+import { unpluginFactory } from './plugin/core'
 
 export default createWebpackPlugin(unpluginFactory)
