+++ website/package.json
@@ -38,6 +38,7 @@
     "xstate": "^5.6.1"
   },
   "devDependencies": {
+    "@pandabox/unplugin-panda-macro": "workspace:^",
     "@pandacss/dev": "^0.30.1",
     "@pandacss/preset-base": "^0.30.1",
     "@park-ui/panda-preset": "^0.33.0",
